# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autobody Shop API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Client {
  id: number;
  name: string;
  email: string;
  phone: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface ClientInput {
  name: string;
  email: string;
  phone: string;
  address?: string;
  notes?: string;
}

export interface ClientUpdate {
  name?: string;
  email?: string;
  phone?: string;
  address?: string;
  notes?: string;
}

export interface Vehicle {
  id: number;
  clientId: number;
  make: string;
  model: string;
  year: number;
  licensePlate: string;
  /** @nullable */
  vin?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  mileage?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface VehicleInput {
  clientId: number;
  make: string;
  model: string;
  year: number;
  licensePlate: string;
  vin?: string;
  color?: string;
  mileage?: number;
  notes?: string;
}

export interface VehicleUpdate {
  make?: string;
  model?: string;
  year?: number;
  licensePlate?: string;
  vin?: string;
  color?: string;
  mileage?: number;
  notes?: string;
}

export interface VehiclePhoto {
  id: number;
  vehicleId: number;
  url: string;
  /** @nullable */
  caption: string | null;
  createdAt: string;
}

export interface VehiclePhotoInput {
  url: string;
  caption?: string;
}

export type CheckinStatus = typeof CheckinStatus[keyof typeof CheckinStatus];


export const CheckinStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  waiting_parts: 'waiting_parts',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Checkin {
  id: number;
  clientId: number;
  vehicleId: number;
  status: CheckinStatus;
  description: string;
  /** @nullable */
  estimatedCost?: number | null;
  droppedOffAt: string;
  /** @nullable */
  estimatedCompletionDate?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface CheckinInput {
  clientId: number;
  vehicleId: number;
  description: string;
  estimatedCost?: number;
  droppedOffAt: string;
  estimatedCompletionDate?: string;
  notes?: string;
}

export type CheckinUpdateStatus = typeof CheckinUpdateStatus[keyof typeof CheckinUpdateStatus];


export const CheckinUpdateStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  waiting_parts: 'waiting_parts',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface CheckinUpdate {
  status?: CheckinUpdateStatus;
  description?: string;
  estimatedCost?: number;
  estimatedCompletionDate?: string;
  completedAt?: string;
  notes?: string;
}

export interface RepairUpdate {
  id: number;
  checkinId: number;
  message: string;
  /** @nullable */
  status?: string | null;
  createdAt: string;
}

export type RepairUpdateInputStatus = typeof RepairUpdateInputStatus[keyof typeof RepairUpdateInputStatus];


export const RepairUpdateInputStatus = {
  pending: 'pending',
  in_progress: 'in_progress',
  waiting_parts: 'waiting_parts',
  ready: 'ready',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface RepairUpdateInput {
  message: string;
  status?: RepairUpdateInputStatus;
}

export interface Part {
  id: number;
  name: string;
  partNumber: string;
  category: string;
  /** @nullable */
  description?: string | null;
  price: number;
  quantity: number;
  /** @nullable */
  supplier?: string | null;
  createdAt: string;
}

export interface PartInput {
  name: string;
  partNumber: string;
  category: string;
  description?: string;
  price: number;
  quantity: number;
  supplier?: string;
}

export interface PartUpdate {
  name?: string;
  partNumber?: string;
  category?: string;
  description?: string;
  price?: number;
  quantity?: number;
  supplier?: string;
}

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  paid: 'paid',
  partial: 'partial',
  refunded: 'refunded',
} as const;

/**
 * @nullable
 */
export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod] | null;


export const PaymentMethod = {
  cash: 'cash',
  card: 'card',
  check: 'check',
  financing: 'financing',
} as const;

export interface Payment {
  id: number;
  clientId: number;
  checkinId: number;
  amount: number;
  status: PaymentStatus;
  /** @nullable */
  method: PaymentMethod;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  paidAt?: string | null;
  createdAt: string;
}

export type PaymentInputStatus = typeof PaymentInputStatus[keyof typeof PaymentInputStatus];


export const PaymentInputStatus = {
  pending: 'pending',
  paid: 'paid',
  partial: 'partial',
  refunded: 'refunded',
} as const;

export type PaymentInputMethod = typeof PaymentInputMethod[keyof typeof PaymentInputMethod];


export const PaymentInputMethod = {
  cash: 'cash',
  card: 'card',
  check: 'check',
  financing: 'financing',
} as const;

export interface PaymentInput {
  clientId: number;
  checkinId: number;
  amount: number;
  status: PaymentInputStatus;
  method?: PaymentInputMethod;
  notes?: string;
  paidAt?: string;
}

export type PaymentUpdateStatus = typeof PaymentUpdateStatus[keyof typeof PaymentUpdateStatus];


export const PaymentUpdateStatus = {
  pending: 'pending',
  paid: 'paid',
  partial: 'partial',
  refunded: 'refunded',
} as const;

export type PaymentUpdateMethod = typeof PaymentUpdateMethod[keyof typeof PaymentUpdateMethod];


export const PaymentUpdateMethod = {
  cash: 'cash',
  card: 'card',
  check: 'check',
  financing: 'financing',
} as const;

export interface PaymentUpdate {
  amount?: number;
  status?: PaymentUpdateStatus;
  method?: PaymentUpdateMethod;
  notes?: string;
  paidAt?: string;
}

export interface DashboardSummary {
  totalClients: number;
  totalVehicles: number;
  activeCheckins: number;
  completedThisMonth: number;
  revenue: number;
  pendingPayments: number;
  partsInStock: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  checkin: 'checkin',
  update: 'update',
  payment: 'payment',
  completion: 'completion',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  title: string;
  description: string;
  timestamp: string;
}

export type ListClientsParams = {
search?: string;
};

export type ListVehiclesParams = {
clientId?: number;
search?: string;
};

export type ListCheckinsParams = {
status?: string;
clientId?: number;
vehicleId?: number;
};

export type ListPartsParams = {
search?: string;
category?: string;
};

export type ListPaymentsParams = {
clientId?: number;
checkinId?: number;
status?: string;
};