# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Autobody Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Vehicles ---
...
  /checkins:
    get:
      operationId: listCheckins
      tags: [checkins]
      summary: List all check-ins
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
        - name: vehicleId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of check-ins
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Checkin"
    post:
      operationId: createCheckin
      tags: [checkins]
      summary: Create a check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"
  /checkins/{id}:
    get:
      operationId: getCheckin
      tags: [checkins]
      summary: Get a check-in
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Check-in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"
        "404":
          description: Not found
    patch:
      operationId: updateCheckin
      tags: [checkins]
      summary: Update a check-in
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckinUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Checkin"

  # --- Repair Updates ---
  /checkins/{id}/updates:
    get:
      operationId: listRepairUpdates
      tags: [repairs]
      summary: List repair updates for a check-in
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Repair updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RepairUpdate"
    post:
      operationId: addRepairUpdate
      tags: [repairs]
      summary: Add a repair update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepairUpdateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepairUpdate"

  # --- Parts Catalog ---
  /parts:
    get:
      operationId: listParts
      tags: [parts]
      summary: List all parts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Parts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Part"
    post:
      operationId: createPart
      tags: [parts]
      summary: Create a part
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
  /parts/{id}:
    get:
      operationId: getPart
      tags: [parts]
      summary: Get a part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Part
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
        "404":
          description: Not found
    patch:
      operationId: updatePart
      tags: [parts]
      summary: Update a part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Part"
    delete:
      operationId: deletePart
      tags: [parts]
      summary: Delete a part
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Payments ---
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, email, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name, email, phone]
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        notes:
          type: string
...
    Checkin:
      type: object
      required: [id, clientId, vehicleId, status, description, droppedOffAt, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        vehicleId:
          type: integer
        status:
          type: string
          enum: [pending, in_progress, waiting_parts, ready, completed, cancelled]
        description:
          type: string
        estimatedCost:
          type: ["number", "null"]
        droppedOffAt:
          type: string
        estimatedCompletionDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    CheckinInput:
      type: object
      required: [clientId, vehicleId, description, droppedOffAt]
      properties:
        clientId:
          type: integer
        vehicleId:
          type: integer
        description:
          type: string
        estimatedCost:
          type: number
        droppedOffAt:
          type: string
        estimatedCompletionDate:
          type: string
        notes:
          type: string
    CheckinUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, in_progress, waiting_parts, ready, completed, cancelled]
        description:
          type: string
        estimatedCost:
          type: number
        estimatedCompletionDate:
          type: string
        completedAt:
          type: string
        notes:
          type: string
    RepairUpdate:
      type: object
      required: [id, checkinId, message, createdAt]
      properties:
        id:
          type: integer
        checkinId:
          type: integer
        message:
          type: string
        status:
          type: ["string", "null"]
        createdAt:
          type: string
    RepairUpdateInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        status:
          type: string
          enum: [pending, in_progress, waiting_parts, ready, completed, cancelled]
    Part:
      type: object
      required: [id, name, partNumber, category, price, quantity, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        partNumber:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        quantity:
          type: integer
        supplier:
          type: ["string", "null"]
        createdAt:
          type: string
    PartInput:
      type: object
      required: [name, partNumber, category, price, quantity]
      properties:
        name:
          type: string
        partNumber:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
        supplier:
          type: string
    PartUpdate:
      type: object
      properties:
        name:
          type: string
        partNumber:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
        quantity:
          type: integer
        supplier:
          type: string
...