# lib/api-zod/src/generated/types/partUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autobody Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface PartUpdate {
  name?: string;
  partNumber?: string;
  category?: string;
  description?: string;
  price?: number;
  quantity?: number;
  supplier?: string;
}