# lib/api-zod/src/generated/types/repairUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autobody Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface RepairUpdate {
  id: number;
  checkinId: number;
  message: string;
  /** @nullable */
  status?: string | null;
  createdAt: string;
}