# lib/api-zod/src/generated/types/repairUpdateInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Autobody Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { RepairUpdateInputStatus } from './repairUpdateInputStatus';

export interface RepairUpdateInput {
  message: string;
  status?: RepairUpdateInputStatus;
}